Update config.ini
The lower case 'y' is not properly recognized by the Marlin print controller firmware, causing a full XYZ homing to occur during the end script.  This results in the nozzle crashing into the newly printed part on the bed, as well as a collection of really nasty sounds from the printer.